# Campus-AI-main/lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * CampusSphere AI Campus API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiError {
  error: string;
}

export interface Student {
  id: number;
  name: string;
  role: string;
  specialty: string;
  avatarColor: string;
  avatarEmoji: string;
  positionX: number;
  positionY: number;
  personality: string;
  isOnline: boolean;
  createdAt: string;
}

export interface StudentInput {
  name: string;
  role: string;
  specialty: string;
  avatarColor: string;
  avatarEmoji: string;
  positionX: number;
  positionY: number;
  personality: string;
  isOnline?: boolean;
}

export interface StudentUpdate {
  name?: string;
  role?: string;
  specialty?: string;
  avatarColor?: string;
  avatarEmoji?: string;
  positionX?: number;
  positionY?: number;
  personality?: string;
  isOnline?: boolean;
}

export interface Conversation {
  id: number;
  studentId: number;
  title: string;
  createdAt: string;
  updatedAt: string;
}

export interface ConversationInput {
  studentId: number;
  title?: string;
}

export interface Message {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface ConversationWithMessages {
  id: number;
  studentId: number;
  title: string;
  createdAt: string;
  updatedAt: string;
  messages: Message[];
}

export interface MessageInput {
  content: string;
}

export interface MemoryEntry {
  id: number;
  content: string;
  tags: string[];
  /** @nullable */
  studentId: number | null;
  createdAt: string;
}

export interface MemoryEntryInput {
  content: string;
  tags?: string[];
  /** @nullable */
  studentId?: number | null;
}

export type MemorySearchItemMetadata = { [key: string]: unknown };

export interface MemorySearchItem {
  id: string;
  content: string;
  score: number;
  tags: string[];
  metadata?: MemorySearchItemMetadata;
}

export interface MemorySearchResult {
  query: string;
  results: MemorySearchItem[];
}

export type GraphNodeMetadata = { [key: string]: unknown };

export interface GraphNode {
  id: string;
  label: string;
  type: string;
  metadata?: GraphNodeMetadata;
}

export interface GraphEdge {
  id: string;
  source: string;
  target: string;
  label: string;
}

export interface KnowledgeGraph {
  nodes: GraphNode[];
  edges: GraphEdge[];
}

export interface TimelineEvent {
  id: string;
  type: string;
  title: string;
  description: string;
  timestamp: string;
  /** @nullable */
  studentId?: number | null;
  /** @nullable */
  studentName?: string | null;
}

export interface AdminStats {
  totalStudents: number;
  totalConversations: number;
  totalMessages: number;
  totalMemoryEntries: number;
  onlineStudents: number;
  activeToday: number;
}

export interface ActivityEvent {
  id: string;
  type: string;
  description: string;
  timestamp: string;
  /** @nullable */
  studentId?: number | null;
  /** @nullable */
  studentName?: string | null;
}

export type ListConversationsParams = {
studentId?: number;
};

export type SearchMemoryParams = {
query: string;
studentId?: number;
};

export type GetMemoryTimelineParams = {
studentId?: number;
};