# Campus-AI-main/lib/api-zod/src/generated/types/conversationWithMessages.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * CampusSphere AI Campus API
 * OpenAPI spec version: 0.1.0
 */
import type { Message } from './message';

export interface ConversationWithMessages {
  id: number;
  studentId: number;
  title: string;
  createdAt: Date;
  updatedAt: Date;
  messages: Message[];
}