# Campus-AI-main/lib/api-zod/src/generated/types/graphEdge.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * CampusSphere AI Campus API
 * OpenAPI spec version: 0.1.0
 */

export interface GraphEdge {
  id: string;
  source: string;
  target: string;
  label: string;
}