# Campus-AI-main/lib/api-zod/src/generated/types/knowledgeGraph.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * CampusSphere AI Campus API
 * OpenAPI spec version: 0.1.0
 */
import type { GraphEdge } from './graphEdge';
import type { GraphNode } from './graphNode';

export interface KnowledgeGraph {
  nodes: GraphNode[];
  edges: GraphEdge[];
}