# Campus-AI-main/lib/api-zod/src/generated/types/memoryEntry.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * CampusSphere AI Campus API
 * OpenAPI spec version: 0.1.0
 */

export interface MemoryEntry {
  id: number;
  content: string;
  tags: string[];
  /** @nullable */
  studentId: number | null;
  createdAt: Date;
}