# Campus-AI-main/lib/api-zod/src/generated/types/memoryEntryInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * CampusSphere AI Campus API
 * OpenAPI spec version: 0.1.0
 */

export interface MemoryEntryInput {
  content: string;
  tags?: string[];
  /** @nullable */
  studentId?: number | null;
}