# Campus-AI-main/lib/api-zod/src/generated/types/memorySearchResult.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * CampusSphere AI Campus API
 * OpenAPI spec version: 0.1.0
 */
import type { MemorySearchItem } from './memorySearchItem';

export interface MemorySearchResult {
  query: string;
  results: MemorySearchItem[];
}