# Campus-AI-main/lib/api-zod/src/generated/types/student.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * CampusSphere AI Campus API
 * OpenAPI spec version: 0.1.0
 */

export interface Student {
  id: number;
  name: string;
  role: string;
  specialty: string;
  avatarColor: string;
  avatarEmoji: string;
  positionX: number;
  positionY: number;
  personality: string;
  isOnline: boolean;
  createdAt: Date;
}